添加调查和对话功能

+++ Assets/Scrpit/Bean/MVC/TextInfoBean.cs
@@ -12,5 +12,5 @@ public class TextInfoBean : BaseBean
     //文本内容
     public string content;
     //文本顺序
-    public long order;
+    public int order;
 }
